pyproject.toml
[tool.poetry]
name = "arabic-dialects-qna-llm"
version = "0.1.0"
description = ""
authors = ["Kirill Chirkunov <kirill.chirkunov@gmail.com>"]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
anthropic = "^0.39.0"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
jupyter = "^1.1.1"
tqdm = "^4.66.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
